Basic builder implemented for MainWindow
The easiest approach for now is to do our configuration in builder
objects. I am planning on using Gson to write/read a builder object
from disk as a configuation file.

I would like to store our inverted tree's in a better format however
i.e a database or something similar so we don't have to reindex on
startup